+++ stdlib_windows.go
@@ -4,7 +4,7 @@ import "unsafe"
 //#include <stdlib.h>
 import "C"
 
-func stdlibMalloc(n int) unsafe.Pointer {
+func stdlibMalloc(n uintptr) unsafe.Pointer {
 	return C.malloc(C.ulonglong(n))
 }
 
@@ -12,7 +12,7 @@ func stdlibCalloc(n int, sizeof uintptr) unsafe.Pointer {
 	return C.calloc(C.ulonglong(n), C.ulonglong(sizeof))
 }
 
-func stdlibRealloc(p unsafe.Pointer, n int) unsafe.Pointer {
+func stdlibRealloc(p unsafe.Pointer, n uintptr) unsafe.Pointer {
 	return C.realloc(p, C.ulonglong(n))
 }
